Remove Service folder, change all CRUD to Controller
