--- app/src/main/java/com/example/homeactivity/Services/NotificationService.java
@@ -1,4 +1,0 @@
-package com.example.homeactivity.Services;
-
-public class NotificationService {
-}
